modified walkerObject.py so that in walkUniform(), positionList contains a list of integers, where before it contained a list of arrays which each contained a single integer (This was making matplotlibs histogram break, with each bin containing exactly 1 element, plotting what should have been the x axis on the y axis and vise versa, and making each bin a different color, and making the total number of bins equal to the total number of runs, even when manually setting the number of bins to a lower or higher number. I mostly understand why this was happening.)
Also changed walkerObject.py so that its position and start position are stored as numbers rather than a vector. Additionally removed import of vpython because it wasnt being used and made using the debugger a pain.

Added explanations for some variables in main_randomWalk.py, added ability to plot a histogram of the average position of each run for 'numberOfRuns' runs.